+++ src/india_api/internal/service/__init__.py
@@ -1,3 +1,4 @@
-from .server import server, get_db_client
+from .server import server
+from .regions import get_db_client
 
 __all__ = ["server", "get_db_client"]
