+++ src/india_api/internal/service/constants.py
@@ -1,0 +1,3 @@
+import pytz
+
+local_tz = pytz.timezone("Asia/Kolkata")
